🗑️ Chore : 소나 테스트 통과하기 위한 커밋 추가
